+++ app/domain/authentication/authenticate.rb
@@ -83,6 +83,7 @@ module Authentication
           authenticator_name: authenticator_name,
           service:            webservice,
           role:               role,
+          client_ip:          origin,
           success:            true,
           error_message:      nil
         )
@@ -95,6 +96,7 @@ module Authentication
           authenticator_name: authenticator_name,
           service:            webservice,
           role:               role,
+          client_ip:          origin,
           success:            false,
           error_message:      err.message
         )