+++ app/domain/authentication/authn_k8s/inject_client_cert.rb
@@ -21,7 +21,7 @@ module Authentication
         validate_pod_request:   ValidatePodRequest.new,
         audit_log:              ::Audit.logger
       },
-      inputs: %i(conjur_account service_id csr host_id_prefix)
+      inputs: %i(conjur_account service_id csr host_id_prefix origin)
     ) do
 
       # :reek:TooManyStatements
@@ -196,6 +196,7 @@ module Authentication
             authenticator_name: KUBERNETES_AUTHENTICATOR_NAME,
             service: webservice,
             role: host,
+            client_ip: @origin,
             success: true,
             error_message: nil
           )
@@ -208,6 +209,7 @@ module Authentication
             authenticator_name: KUBERNETES_AUTHENTICATOR_NAME,
             service: webservice,
             role: host,
+            client_ip: @origin,
             success: false,
             error_message: err.message
           )