+++ app/domain/authentication/authn_oidc/authenticator.rb
@@ -141,6 +141,7 @@ module Authentication
             authenticator_name: authenticator_name,
             service:            webservice,
             role:               role,
+            client_ip:          origin,
             success:            true,
             error_message:      nil
           )
@@ -153,6 +154,7 @@ module Authentication
             authenticator_name: authenticator_name,
             service:            webservice,
             role:               role,
+            client_ip:          origin,
             success:            false,
             error_message:      err.message
           )