+++ app/domain/authentication/login.rb
@@ -21,7 +21,7 @@ module Authentication
     extend Forwardable
     def_delegators(
       :@authenticator_input, :authenticator_name, :account, :username,
-      :webservice, :role
+      :webservice, :role, :origin
     )
 
     def call
@@ -77,6 +77,7 @@ module Authentication
           authenticator_name: authenticator_name,
           service:            webservice,
           role:               role,
+          client_ip:          origin,
           success:            true,
           error_message:      nil
         )
@@ -89,6 +90,7 @@ module Authentication
           authenticator_name: authenticator_name,
           service:            webservice,
           role:               role,
+          client_ip:          origin,
           success:            false,
           error_message:      err.message
         )