+++ spec/app/domain/authentication/validate_status_spec.rb
@@ -108,6 +108,9 @@ RSpec.describe Authentication::ValidateStatus do
       allow(status_input).to receive(:username)
                                .and_return(test_user_id)
 
+      allow(status_input).to receive(:origin)
+                               .and_return(test_origin)
+
       allow(status_input).to receive(:role)
                                .and_return(mock_role_class)
     end